// frontend/src/api/generated/models/product-page-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ProductDetail } from './product-detail';

/**
 * 
 * @export
 * @interface ProductPageResponse
 */
export interface ProductPageResponse {
    /**
     * 
     * @type {Array<ProductDetail>}
     * @memberof ProductPageResponse
     */
    'content': Array<ProductDetail>;
    /**
     * 
     * @type {number}
     * @memberof ProductPageResponse
     */
    'page': number;
    /**
     * 
     * @type {number}
     * @memberof ProductPageResponse
     */
    'size': number;
    /**
     * 
     * @type {number}
     * @memberof ProductPageResponse
     */
    'totalElements': number;
    /**
     * 
     * @type {number}
     * @memberof ProductPageResponse
     */
    'totalPages': number;
}