// frontend/src/api/generated/models/storage-condition.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 保管条件（常温・冷蔵・冷凍）
 * @export
 * @enum {string}
 */

export const StorageCondition = {
    Ambient: 'AMBIENT',
    Refrigerated: 'REFRIGERATED',
    Frozen: 'FROZEN'
} as const;

export type StorageCondition = typeof StorageCondition[keyof typeof StorageCondition];